requirements.txt
boto3==1.28.57
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0